GitButler Integration Commit
This is an integration commit for the virtual branches that GitButler is tracking.

Due to GitButler managing multiple virtual branches, you cannot switch back and
forth between git branches and virtual branches easily. 

If you switch to another branch, GitButler will need to be reinitialized.
If you commit on this branch, GitButler will throw it away.

Here are the branches that are currently applied:
 - testing (refs/gitbutler/testing)
   branch head: ba646f4f4461511721b970726ca91556fabee5f4
   - src/components/Navbar.vue
   - .gitignore
   - src/assets/xel-logo.png
   - src/assets/icons/3-bar.svg
   - src/assets/v.jpg
   - tailwind.config.js

Your previous branch was: refs/heads/testing

The sha for that commit was: 4facbc3bab29d353d91ef12f61337020ffd33546

For more information about what we're doing here, check out our docs:
https://docs.gitbutler.com/features/virtual-branches/integration-branch